% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vina.R
\docType{package}
\name{autodockr-package}
\alias{autodockr-package}
\alias{autodockr}
\title{R interface for Autodock Vina}
\description{
This is an R interface to allow calling Autodock Vina from whithin R. The package compile Vina
and its dependencies upon installation.
}
\author{
Ahmed Mohamed \email{mohamed@kuicr.kyoto-u.ac.jp}
}
